+++ package.json
@@ -9,6 +9,7 @@
     "lint": "next lint"
   },
   "dependencies": {
+    "@reduxjs/toolkit": "^1.9.5",
     "@types/aos": "^3.0.4",
     "@types/node": "^20.5.6",
     "@types/react": "^18.2.21",
@@ -24,7 +25,6 @@
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-icons": "^4.10.1",
-    "react-intersection-observer": "^9.5.2",
     "react-redux": "^8.1.2",
     "react-slick": "^0.29.0",
     "slick-carousel": "^1.8.1",
