+++ src/app/services/page.tsx
@@ -1,11 +1,9 @@
-import { OurServices } from "./OurServices";
-import { ServiceHero } from "./ServiceHero";
+import Services from "@/components/Services";
 
 export default function ServicesPage() {
   return (
     <main>
-      <ServiceHero />
-      <OurServices />
+      <Services />
     </main>
   );
 }
