--- src/app/works/works.ts
@@ -1,101 +1,0 @@
-export const worksData = [
-  {
-    title: "Nature UI",
-    description:
-      "Nature UI is a Modular, Composable and Extensible React based component library built upon tailwindcss and emotion",
-    imageUrl:
-      "https://res.cloudinary.com/dnature/image/upload/c_scale,w_0.8/v1614461932/works/nature-ui_jleesu.png",
-    githubUrl: "https://github.com/nature-ui/nature-ui",
-    websiteUrl: "https://nature-ui.com",
-    tags: ["#open-source", "#react", "#typescript"],
-  },
-  {
-    title: "Cryptocompare dashboard",
-    description:
-      "Crypto compare is a real-time crypto dashboard built with React and Typescript.",
-    imageUrl:
-      "https://res.cloudinary.com/dnature/image/upload/c_scale,w_0.8/v1625513934/works/crypto-tracker_dnfvd5.png",
-    githubUrl: "https://github.com/DNature/bitcoin-tracker",
-    websiteUrl: "https://bitcoin-tracker-ashy.vercel.app",
-    tags: ["#crypto", "#react", "#typescript"],
-  },
-  {
-    title: "TradingView Binance",
-    description:
-      "Tradingview charting library & Binance implementation with React.",
-    imageUrl:
-      "https://res.cloudinary.com/dnature/image/upload/c_scale,w_0.8/v1627514672/works/yv_ruihgt.png",
-    githubUrl: "https://github.com/DNature/tv-next-binance",
-    websiteUrl: "https://tv-next-binance.vercel.app",
-    tags: ["#crypto", "#nextjs", "#typescript"],
-  },
-  {
-    title: "CombatCovid",
-    description:
-      "Combat covid with hardware Search for well documented designs 🔍 and document new solutions.",
-    imageUrl:
-      "https://res.cloudinary.com/dnature/image/upload/c_scale,w_0.8/v1614461947/works/combat-covid_qxwl74.png",
-    githubUrl: "http://github.com/CombatCovid",
-    websiteUrl: "http://combatcovid.equipment/",
-    tags: ["#open-source", "#vue"],
-  },
-  {
-    title: "TradingView Chart widget",
-    description: "Tradingview charting library widget with React.",
-    imageUrl:
-      "https://res.cloudinary.com/dnature/image/upload/c_scale,w_0.8/v1627514295/works/tv2_mtrain.png",
-    githubUrl: "https://github.com/DNature/tv-react-v2",
-    websiteUrl: "https://v2-tv-react.netlify.app",
-    tags: ["#crypto", "#react", "#typescript"],
-  },
-  {
-    title: "Hardocs",
-    description:
-      "Hardocs is a cross Platform desktop app for project documentation Designed with hardware documentation in mind.",
-    imageUrl:
-      "https://res.cloudinary.com/dnature/image/upload/c_scale,w_0.8/v1614461924/works/hardocs_xxcozk.png",
-    githubUrl: "https://github.com/hardocs/desktop-app",
-    websiteUrl: "https://hardocs.github.io",
-    tags: ["#open-source", "#electron", "#vue"],
-  },
-  {
-    title: "CRWN Clothing",
-    description:
-      "CRWN clothing is an imaginary e-commerce clothing store. Some features includes payment with stripe, redux for global state management, firebase and lots more.",
-    imageUrl:
-      "https://res.cloudinary.com/dnature/image/upload/c_scale,w_0.8/v1614461940/works/crwn-clothing_eqs9nt.png",
-    githubUrl: "http://github.com/DNature/crwn-clothing-e-commerce",
-    websiteUrl: "http://crwn-clothing-navy-ten.now.sh/",
-    tags: ["#react", "#e-commerce", "#firebase"],
-  },
-  {
-    title: "Transferwise Clone",
-    description:
-      "I cloned the homepage of transferwise.com just for fun. I attempted the design and most functionalities present on the website.",
-    imageUrl:
-      "https://res.cloudinary.com/dnature/image/upload/c_scale,w_0.8/v1614461944/works/transferwise_ggpif9.png",
-    githubUrl: "http://github.com/DNature/transferwise",
-    websiteUrl: "https://transferwise.netlify.app/",
-    tags: ["#react", "#ui"],
-  },
-  {
-    title: "Movie Time",
-    description:
-      "A web app for booking movie tickets from a fictional cinema. It features movie trailers, rating, description, online payment and email notifications.!",
-    imageUrl:
-      "https://res.cloudinary.com/dnature/image/upload/c_scale,w_0.8/v1614465254/works/movies-time_szkzpw.png",
-    githubUrl: "http://github.com/DNature/movietime",
-    websiteUrl: "http://movietimesng.herokuapp.com/",
-    tags: ["#react", "#ui"],
-  },
-  {
-    title: "Cleanfits",
-    description:
-      "Cleanfits is an imaginary laundry and dry cleaning agency. It lets users book cleaning appointments, track delivery and pay online.!",
-    imageUrl:
-      "https://res.cloudinary.com/dnature/image/upload/c_scale,w_0.8/v1614509870/works/clean-fits_x1m6uf.png",
-    githubUrl: "http://github.com/DNature/cleanfits_original",
-    websiteUrl: "http://cleanfits.herokuapp.com/",
-    tags: ["#react", "#ui"],
-  },
-];
