+++ src/components/Home/index.tsx
@@ -2,13 +2,13 @@ import { Contact } from "./components/Contact";
 import { Hero } from "./components/Hero";
 import { Projects } from "./components/Projects";
 import { Testimonial } from "./components/Testimonial";
-import { OurProcess } from "./components/WhatWeDo";
+import { WhatWeDo } from "./components/WhatWeDo";
 
 export default function Home() {
   return (
     <main>
       <Hero />
-      <OurProcess />
+      <WhatWeDo />
       <Projects />
       <Testimonial />
       <Contact />
