+++ src/components/Services/index.tsx
@@ -1,0 +1,11 @@
+import { OurServices } from "./components/OurServices";
+import { ServiceHero } from "./components/ServiceHero";
+
+export default function Services() {
+  return (
+    <main>
+      <ServiceHero />
+      <OurServices />
+    </main>
+  );
+}
