+++ src/redux/provider.tsx
@@ -1,0 +1,9 @@
+"use client";
+
+import { Provider } from "react-redux";
+import { store } from "./store";
+import { ReactNode } from "react";
+
+export function ReduxProvider({ children }: { children: ReactNode }) {
+  return <Provider store={store}>{children}</Provider>;
+}
