+++ src/redux/store.ts
@@ -1,0 +1,12 @@
+import { configureStore } from "@reduxjs/toolkit";
+import authReducer from "./features/authSlice";
+import { TypedUseSelectorHook, useSelector } from "react-redux";
+export const store = configureStore({
+  reducer: {
+    authReducer,
+  },
+});
+
+export type RootState = ReturnType<typeof store.getState>;
+export type AppDispatch = typeof store.dispatch;
+export const useAppSelector: TypedUseSelectorHook<RootState> = useSelector;
