--- src/utils/getNchars.tsx
@@ -1,8 +1,0 @@
-export const getNChars = (val: string, n = 80): string => {
-  const sortNChars = val.substring(0, n);
-  if (val.length <= sortNChars.length) {
-    return val;
-  }
-
-  return sortNChars + "...";
-};
